Revert "Revert "ModuleDependenciesCollector""
This reverts commit 32ff38801f23579d1367fb71458d8a88729a5cb2.
